[web] move tests to text module from example, use wasm-pack test framework
