+++ piet-web/Cargo.toml
@@ -23,3 +23,11 @@ xi-unicode = "0.2.0"
 version = "0.3.36"
 features = ["Window", "CanvasGradient", "CanvasRenderingContext2d", "CanvasWindingRule",
     "Document", "Element", "HtmlCanvasElement", "ImageBitmap", "ImageData", "TextMetrics"]
+
+[dev-dependencies]
+wasm-bindgen-test = "0.3.0"
+
+[dev-dependencies.web-sys]
+version = "0.3.36"
+features = ["console", "Window", "CanvasGradient", "CanvasRenderingContext2d", "CanvasWindingRule",
+    "Document", "Element", "HtmlCanvasElement", "ImageBitmap", "ImageData", "TextMetrics"]
